% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{c.delay_dist}
\alias{c.delay_dist}
\title{Combines multiple delay distributions into a new delay distribution}
\usage{
\method{c}{delay_dist}(...)
}
\arguments{
\item{...}{The delay distributions (from calls to \code{\link[=dist_spec]{dist_spec()}}) to combine}
}
\value{
Combined delay distributions (with class \code{\link[=dist_spec]{dist_spec()}}`)
}
\description{
At the moment all this does is concatenate the parameters so that they can
be fed as multiple delay distributions to \code{\link[=epinow]{epinow()}} or \code{\link[=estimate_infections]{estimate_infections()}}.
}
\details{
In the future this might be extended to simplify distributions,  e.g. when
multiple gamma distributions are used.
}
\author{
Sebastian Funk
}
