% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_seeding_time}
\alias{create_seeding_time}
\title{Create seeding time from delays and generation time}
\usage{
create_seeding_time(delays, generation_time)
}
\arguments{
\item{delays}{An object of type \code{delay_dist()}, representing the delays}

\item{generation_time}{An object of type \code{delay_dist()}, representing the
generation_time}
}
\value{
Integer; the seeding time
}
\description{
The seeding time is calculated as the sum of the means of the given
delay distribution, but at least the maximum generation time.
}
\author{
Sebastian Funk
}
